print: Always cast the error value to int
Avoid burden on the caller to cast the value to int. We will get printf
format warning if, for example, the integer type to be passed in is
greater than int, e.g. on x86-64 we have ssize_t != int, this will
generate silly printf format warnings.

Just always cast it to int to clean it up!

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>